% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{package}
\name{horsey-package}
\alias{horsey}
\alias{horsey-package}
\title{horsey: Tools for Analyzing and Exporting Data from the Lichess API}
\description{
Client that accesses the Lichess API. Supports both authenticated and unauthenticated requests.
}
\author{
\strong{Maintainer}: W. Kyle Hamilton \email{kyle.hamilton@gmail.com}

}
\keyword{internal}
