% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_openings.R
\name{plot_opening_counts}
\alias{plot_opening_counts}
\title{Plot most common openings}
\usage{
plot_opening_counts(games, top_n = 10, include_eco = TRUE)
}
\arguments{
\item{games}{A tibble returned by \code{horsey::export_games()}.}

\item{top_n}{Number of openings to show.}

\item{include_eco}{Logical. If \code{TRUE}, prepend ECO code to the label.}
}
\value{
A \code{ggplot2} object.
}
\description{
Builds a horizontal bar chart of the most frequent openings found in
\code{horsey::export_games()} output when \code{opening = TRUE} was requested.
}
