% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results.R
\name{plot_results_over_time}
\alias{plot_results_over_time}
\title{Plot rolling results over time}
\usage{
plot_results_over_time(games, player, by = c("week", "month"))
}
\arguments{
\item{games}{A tibble returned by \code{horsey::export_games()}.}

\item{player}{Lichess username to evaluate from that player's perspective.}

\item{by}{Aggregate by \code{"week"} or \code{"month"}.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot rolling results over time
}
