% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results.R
\name{plot_time_control_results}
\alias{plot_time_control_results}
\title{Plot result mix by time control}
\usage{
plot_time_control_results(games, player)
}
\arguments{
\item{games}{A tibble returned by \code{horsey::export_games()}.}

\item{player}{Lichess username to evaluate from that player's perspective.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot result mix by time control
}
