% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{rating_history}
\alias{rating_history}
\title{Get a user's rating history}
\usage{
rating_history(cl, username)
}
\arguments{
\item{cl}{A \code{horsey_client}.}

\item{username}{A Lichess username.}
}
\value{
A tibble with columns perf, rating, date.
}
\description{
Returns rating history for each game type (e.g., blitz, rapid).
}
