% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{user}
\alias{user}
\title{Get a user's public profile}
\usage{
user(cl, username)
}
\arguments{
\item{cl}{A \code{horsey_client}.}

\item{username}{Lichess username.}
}
\value{
A tibble with selected fields, plus a \code{raw} list-column.
}
\description{
Retrieves a public user profile from the Lichess API.
}
